Refactor DataTable utility imports and reorganize file structure
Update import paths for utility functions in DataTable components to reflect the new utils directory. Remove the now redundant deep-utils and table-config files, enhancing code clarity and maintainability. This change streamlines the DataTable component's structure and improves organization.
