+++ src/components/data-table/data-table.tsx
@@ -33,7 +33,7 @@ import { DataTableToolbar } from "@/components/data-table/toolbar";
 import { Skeleton } from "@/components/ui/skeleton";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { AlertCircle } from "lucide-react";
-import { useTableConfig, TableConfig } from "@/components/data-table/table-config";
+import { useTableConfig, TableConfig } from "@/components/data-table/utils/table-config";
 import { useTableColumnResize } from "@/components/data-table/hooks/use-table-column-resize";
 import { DataTableResizer } from "@/components/data-table/data-table-resizer";
