+++ src/components/data-table/toolbar.tsx
@@ -17,9 +17,9 @@ import { Separator } from "@/components/ui/separator";
 import { CalendarDatePicker } from "@/components/calendar-date-picker";
 import { DataTableViewOptions } from "@/components/data-table/view-options";
 import { DataTableExport } from "./data-export";
-import { resetUrlState } from "@/components/data-table/deep-utils";
+import { resetUrlState } from "@/components/data-table/utils/deep-utils";
 import { parseDateFromUrl } from "@/components/data-table/utils/url-state";
-import { TableConfig } from "@/components/data-table/table-config";
+import { TableConfig } from "@/components/data-table/utils/table-config";
 import { formatDate } from "@/components/data-table/utils/date-formate";
 
 interface DataTableToolbarProps<TData> {
